==> notebooks/01_baseline.ipynb <==
%reload_ext autoreload
%autoreload 2

import jupytools.syspath
jupytools.syspath.add('..')

import catboost as cb
import feather
from IPython.display import FileLink
from dataset import load, load_sample, Subset
from extract_features import extend_with_event_data, prepare, baseline_features

trn_data, trn_target, trn_specs = load(Subset.Train)

[tst_data] = load(Subset.Test)

X_trn = extend_with_event_data(trn_data)
X_trn = prepare(X_trn, baseline_features, targets=trn_target)
X_trn = X_trn.drop_duplicates(subset=['game_session', 'installation_id'], keep='last')
X_trn = X_trn.reset_index(drop=True)
X_trn.to_feather('X_trn.feather')

X_tst = extend_with_event_data(tst_data.iloc[:10000])
X_tst = prepare(X_tst, baseline_features)
X_tst = X_tst.drop_duplicates(subset=['game_session', 'installation_id'], keep='last')
X_tst = X_tst.reset_index(drop=True)
X_tst.to_feather('X_tst.feather')

trn_target[['accuracy_group']].to_feather('y.feather')

X_trn = feather.read_dataframe('X_trn.feather')
X_tst = feather.read_dataframe('X_tst.feather')
y = feather.read_dataframe('y.feather').accuracy_group.values

assert X_trn.shape[0] == y.shape[0]

params = dict(
    loss_function='RMSE',
    custom_metric=['RMSE', 'MAE'],
    eval_metric='RMSE',
    iterations=10000,
    random_seed=1,
    max_depth=6,
    early_stopping_rounds=100,
    task_type='GPU',
    devices='0:1',
    verbose=100
)

X_trn_game_session = X_trn.pop('game_session')
X_trn_installation_id = X_trn.pop('installation_id')
cat_cols = X_trn.select_dtypes(object).columns.tolist()

cv_results = cb.cv(params=params, pool=cb.Pool(X_trn, y, cat_features=cat_cols))

train_params = params.copy()
train_params['iterations'] = 1100
booster = cb.train(params=train_params, pool=cb.Pool(X_trn, y, cat_features=cat_cols))
booster.save_model('catboost_baseline.cbm')

FileLink('catboost_baseline.cbm')

X_tst_game_session = X_tst.pop('game_session')
X_tst_installation_id = X_tst.pop('installation_id')
preds = booster.predict(X_tst)

preds_raw = pd.DataFrame(
    {'game_session': X_tst_game_session,
     'installation_id': X_tst_installation_id,
     'prediction': preds})

submit = (preds_raw
 .groupby('installation_id', as_index=False)
 .mean()
 .assign(accuracy_group=lambda df: df['prediction'].round().astype(int))
 .drop(columns=['prediction']))

submit.to_csv('submit_smoketest.csv', index=False)

FileLink('submit_smoketest.csv')
