==> notebooks/02_low_memory_features_extraction.ipynb <==
%reload_ext autoreload
%autoreload 2

import jupytools.syspath
jupytools.syspath.add('..')

import csv
import os
import catboost as cb
import feather
import pandas as pd
from multiprocessing import cpu_count
from tqdm.auto import tqdm
from basedir import TEST
from extract_features import prepare, prepare_groups, baseline_features, extend_with_event_data
from utils import chunks, parallel

dataset = pd.read_csv(TEST)
dataset['timestamp'] = pd.to_datetime(dataset['timestamp'])
dataset.sort_values(by='timestamp', inplace=True)

group_sizes = dataset.groupby('installation_id', as_index=False).game_session.count()
single_session = group_sizes.query('game_session == 1').installation_id
dataset[dataset.installation_id.isin(single_session)]

def save_installation_groups(dataset, output_dir):
    os.makedirs(output_dir, exist_ok=True)
    groups = dataset.groupby('installation_id')
    with tqdm(total=groups.ngroups) as bar:
        for key, group in groups:
            bar.set_description(key)
            filename = os.path.join(output_dir, key)
            group = group.reset_index(drop=True)
            group.to_feather(filename)
            bar.update(1)
    filenames = [os.path.join(output_dir, fn) for fn in os.listdir(output_dir)]
    return filenames

def prepare_on_disk(items, features, buffer=16, dedup=True,
                    num_workers=cpu_count(),
                    output_dir='/tmp/prepared'):
    
    from uuid import uuid4
    os.makedirs(output_dir, exist_ok=True)
    file_batches = list(chunks(items, buffer))
    prepared_files = []
    for filenames in tqdm(file_batches, desc='Processing files'):
        datasets = parallel(feather.read_dataframe, filenames, num_workers=num_workers)
        datasets = parallel(extend_with_event_data, datasets, num_workers=num_workers)
        prepared = prepare_groups(datasets, features, num_workers=num_workers)
        if dedup:
            prepared = prepared.drop_duplicates(
                subset=['game_session', 'installation_id'],
                keep='last')
            prepared = prepared.reset_index(drop=True)
        filename = os.path.join(output_dir, str(uuid4()))
        prepared.to_feather(filename)
        prepared_files.append(filename)
    return prepared_files

grouped_filenames = save_installation_groups(dataset, '/tmp/installations')

prepared_filenames = prepare_on_disk(grouped_filenames, baseline_features)

def run_model(model_file, data_files):
    booster = cb.CatBoost()
    booster.load_model(model_file)
    for filename in data_files:
