==> notebooks/03_in_memory_feature_extraction.ipynb <==
%matplotlib inline
%reload_ext autoreload
%autoreload 2

import jupytools.syspath
jupytools.syspath.add('..')

import feather
import pandas as pd
from basedir import TRAIN, TEST
from dataset import load, load_sample, Subset
from extract_features import prepare, BaselineWithTimestampFeatures

trn_data, trn_targ, trn_spec = load(Subset.Train)

trn_data = prepare(trn_data, BaselineWithTimestampFeatures(), 
                   targets=trn_targ, pbar=True, dedup=True,
                   num_workers=12)

trn_data.to_feather('X_trn.feather')

tst_data = prepare(tst_data, BaselineWithTimestampFeatures(), pbar=True, dedup=True, num_workers=12)

tst_data.to_feather('X_tst.feather')

trn_targ[['accuracy_group']].to_feather('y.feather')

X_trn = feather.read_dataframe('X_trn.feather')
