==> notebooks/17_te.ipynb <==
# !cp /kaggle/input/data-bowl-2019-external-data/*.py /kaggle/working

%reload_ext autoreload
%autoreload 2
import warnings
import jupytools.syspath
def ignore(*args, **kwargs): pass
warnings.warn = ignore
jupytools.syspath.add('..')

from collections import Counter, OrderedDict
from functools import partial

import numpy as np
import pandas as pd
from IPython.display import display
from sklearn.metrics import cohen_kappa_score
from sklearn.model_selection import GroupKFold
from tqdm.auto import tqdm

import bundle
import features as F
import selection
import utils as U
from dataset import load, load_sample, Subset
from encode import encode
from training import train, inference, submit, EnsembleTrainer, get_default_config
from meta import compute_meta_data
from metric import optimize_rounding_bounds, make_cappa_metric
from normalize import normalize

sample = False
if U.on_kaggle():
    U.log('Loading test set only.')
    tst_data = pd.read_csv('/kaggle/input/data-science-bowl-2019/test.csv')
else:
    if sample:
        U.log('Warning: loading train and test data sample.')
        trn_data, _, _ = load_sample(Subset.Train, 500_000)
        [tst_data] = load_sample(Subset.Test, 500_000)
    else:
        U.log('Loading train and test.')
        trn_data, trn_spec, trn_targ = load(Subset.Train)
        [tst_data] = load(Subset.Test)

transform = U.combine(
    partial(F.add_feature_combinations, pairs=[('title', 'event_code'),
                                               ('title', 'world'),
                                               ('title', 'type'),
                                               ('world', 'type')]),
    partial(F.add_datetime, column='timestamp', prefix='ts'),
)

if U.on_kaggle():
    U.log('Transforming test data only.')
    X_tst = transform(tst_data.copy())
    U.log(X_tst.shape)
else:
    U.log('Transforming train and test data.')
    X_tst = transform(tst_data.copy())
    X_trn = transform(trn_data.copy())
    U.log(X_trn.shape, X_tst.shape)

if U.on_kaggle():
    U.log('Reading pre-computed meta from disk.')
    meta = bundle.meta()
else:
    U.log('Computing meta using train and test datasets.')
    meta = compute_meta_data(X_trn, X_tst)
    U.log('Saving computed meta on disk.')
    bundle.save_meta(meta, 'meta')

extractor = F.FeaturesExtractor([
    F.CountingFeatures(meta),
    F.PerformanceFeatures(meta),
    F.VarietyFeatures(meta),
    F.EventDataFeatures(meta),
    F.FeedbackFeatures(meta),
    F.TimestampFeatures2(meta),
    # F.ZFeatures(meta)
])

algo = F.InMemoryAlgorithm(extractor, meta, num_workers=12)

cat_cols = ['session_title','cnt_most_freq_title', 'cnt_least_freq_title', 
            'cnt_most_freq_world', 'cnt_least_freq_world', 'cnt_most_freq_type', 
            'cnt_least_freq_type', 'cnt_most_freq_title_world', 
            'cnt_least_freq_title_world', 'cnt_most_freq_title_type', 
            'cnt_least_freq_title_type', 'cnt_most_freq_world_type', 
            'cnt_least_freq_world_type']

if U.on_kaggle():
    U.log('Preparing test dataset.')
    X_tst = algo.run(X_tst, test=True)
    encoders = bundle.encoders()
    X_tst, _ = encode(X_tst, cat_cols, encoders=encoders)
else:
    U.log('Preparing train and test datasets.')
    X_trn = algo.run(X_trn)
    X_tst = algo.run(X_tst, test=True)
    X_trn, encoders = encode(X_trn, cat_cols)
    X_tst, _ = encode(X_tst, cat_cols, encoders=encoders)
    bundle.save(encoders, 'encoders')

if U.on_kaggle():
    U.log('Running post-processing on test set only.')
    F.add_user_wise_features(X_tst, meta)
else:
    U.log('Running post-processing on train and test sets.')
    F.add_user_wise_features(X_trn, meta)
    F.add_user_wise_features(X_tst, meta)

te_features = list(np.unique(cat_cols + U.starts_with(X_trn.columns, 'cnt_')))

def encode_multi_target(dataset, feature, target='accuracy_group', smoothing=3.0):
    from category_encoders import TargetEncoder
    mappings = []
    for i in (1, 2, 3):
        target_col = dataset[target].map(lambda x: int(x == i))
        te = TargetEncoder(cols=[feature], smoothing=smoothing)
        te.fit(dataset, target_col)
        mapping = te.mapping[feature]
        mapping = {k-1: v for k, v in mapping.to_dict().items() if k > 0}
        dataset[f'te_{feature}_{i}'] = dataset[feature].map(mapping)
        mappings.append(mapping)
    return feature, mappings

if U.on_kaggle():
    te = bundle.target_encoders()
    for feature, mappings in te.items():
        for i, mapping in enumerate(mappings, 1):
            X_tst[f'te_{feature}_{i}'] = X_tst[feature].map(mapping)
else:
    target_encoders = {}
    for feature in tqdm(te_features, desc='Target Encoder'):
        feature, mappings = encode_multi_target(X_trn, feature=feature)
        target_encoders[feature] = mappings
    for feature, mappings in target_encoders.items():
        for i, mapping in enumerate(mappings, 1):
            X_tst[f'te_{feature}_{i}'] = X_tst[feature].map(mapping)
    bundle.save(target_encoders, 'target_encoders')

for feature, mappings in target_encoders.items():
    for i, mapping in enumerate(mappings, 1):
        X_trn[f'te_{feature}_{i}'] = X_trn[feature].map(mapping)

# if not U.on_kaggle():
#     with pd.option_context('display.max_rows', 2000):
#         display(X_trn.sample(3).T)

if not U.on_kaggle():
    X_trn.to_pickle('/tmp/X_trn.pickle')
    X_tst.to_pickle('/tmp/X_tst.pickle')

if not U.on_kaggle():
    X_trn = pd.read_pickle('/tmp/X_trn.pickle')
    X_tst = pd.read_pickle('/tmp/X_tst.pickle')

# group_col = 'installation_id'
# U.log(f'Normalizing dataset using column for grouping: {group_col}')
# norm_dataset = X_tst if U.on_kaggle() else X_trn
# cnt_cols = U.starts_with(norm_dataset.columns, 'cnt_')
# normalize(norm_dataset, cnt_cols, grouping_key=group_col, method='min-max')

## Features Selection

selector = selection.FeatureSelection(
    rules=[
        ('nonzero', selection.non_zero_rows_and_cols),
    ],
    ignore_cols=[
        'accuracy_group', 
        'installation_id', 
        'game_session'
    ]
)
if U.on_kaggle():
    U.log('Loading relevant features list from disk.')
    features = bundle.features()
else:
    U.log('Deriving relevant features from train dataset.')
    features = selector.select(X_trn)
    bundle.save(features, 'features')

## Submission

bundle.save({
    'rounders': rounders, 
    'ensemble': [result.models for result in ensemble]
}, key='lightgbm')

bundle.package(folder='/home/ck/data/bowl2019/external/')

if U.on_kaggle():
    U.log('Inference on Kaggle.')
    features = bundle.features()
    lightgbm = bundle.load('lightgbm')
    labels = []
    U.log('Running model predictions.')
    for rounder in lightgbm['rounders']:
        U.log(f'Using rounding method: {rounder.method}')
        for models in lightgbm['ensemble']:
            for model in models:
                model_pred = model.predict(X_tst[features])
                model_labels = rounder.predict(model_pred)
                labels.append(model_labels)
    labels = np.column_stack(labels)
    predictions = []
    U.log('Majority vote.')
    for label in labels:
        majority = Counter(label).most_common(1)[0][0]
        predictions.append(majority)
    U.log('Saving predictions on disk.')
    filename = submit(predictions)
    submit_df = pd.read_csv(filename)
    U.log('First 20 submission rows:')
    display(submit_df.head(20))
            
    #features = bundle.features()
    #bounds = bundle.bounds()
    #predicted = inference(X_tst, features, bounds=bounds, model=algo, version=version)
    #U.log('Saving predictions on disk.')
    #filename = submit(predicted)
    #submit_df = pd.read_csv(filename)
    #U.log('First 20 submission rows:')
    #display(submit_df.head(20))

rounders = [Rounder(method='dist'), Rounder(method='optimal')]
labels = []
for rounder in rounders:
    for i, result in enumerate(ensemble, 1):
        rounder.fit(X_trn['accuracy_group'], result.oof)
        for j, model in enumerate(result.models, 1):
            U.log(f'Rounder: {rounder.method}, ensemble #{i}, model #{j}')
            model_pred = model.predict(X_tst[features])
            model_labels = rounder.predict(model_pred)
            labels.append(model_labels)

from rounding import Rounder

def seed_all(value):
    import random
    random.seed(value)
    np.random.seed(value)

features = bundle.features()

ensemble = []
U.log('Start training an ensemble of models with various random seeds.')
trainer = EnsembleTrainer(algo='lightgbm', eval_metric='rmse')
fold = GroupKFold(n_splits=5)
for i in range(5):
    seed = np.random.randint(0, 1001)
    U.log('=' * 80)
    U.log(f'Training model with random seed: {seed}')
    seed_all(seed)
    config = get_default_config('lightgbm')
    U.set_nested(config, 'model_params.seed', seed)
    U.set_nested(config, 'model_params.feature_fraction', 0.4)
    U.set_nested(config, 'model_params.bagging_fraction', 0.6)
    U.set_nested(config, 'model_params.bagging_freq', 1)
    U.set_nested(config, 'model_params.learning_rate', 0.003)
    U.set_nested(config, 'model_params.n_estimators', 10000)
    U.set_nested(config, 'fit_params.verbose', 250)
    U.set_nested(config, 'fit_params.early_stopping_rounds', 250)
    result = trainer.train(X_trn, features=features, fold=fold, config=config)
    ensemble.append(result)

rounders = [Rounder(method='dist'), Rounder(method='optimal')]
labels = []
for rounder in rounders:
    for i, result in enumerate(ensemble, 1):
        rounder.fit(X_trn['accuracy_group'], result.oof)
        for j, model in enumerate(result.models, 1):
            U.log(f'Rounder: {rounder.method}, ensemble #{i}, model #{j}')
            model_pred = model.predict(X_tst[features])
            model_labels = rounder.predict(model_pred)
            labels.append(model_labels)

bundle.save({
    'rounders': rounders, 
    'ensemble': [result.models for result in ensemble]
}, key='lightgbm')

from multiprocessing import Queue, Process

ensemble = []
U.log('Start training an ensemble of models with various random seeds.')
trainer = EnsembleTrainer(algo='xgboost', eval_metric='rmse')
fold = GroupKFold(n_splits=5)
for i in range(5):
    seed = np.random.randint(0, 1001)
    U.log('=' * 80)
    U.log(f'Training model with random seed: {seed}')
    seed_all(seed)
    config = get_default_config('xgboost')
    U.set_nested(config, 'model_params.seed', seed)
    U.set_nested(config, 'model_params.max_depth', 6)
    U.set_nested(config, 'model_params.subsample', 0.5)
    U.set_nested(config, 'model_params.colsample_bytree', 0.8)
    U.set_nested(config, 'model_params.colsample_bylevel', 0.6)
    U.set_nested(config, 'model_params.colsample_bynode', 0.5)
    U.set_nested(config, 'model_params.learning_rate', 0.003)
    U.set_nested(config, 'model_params.n_estimators', 5000)
    U.set_nested(config, 'model_params.lambda', 1)
    U.set_nested(config, 'model_params.alpha', 1)
    U.set_nested(config, 'model_params.objective', 'reg:squarederror')
    # U.set_nested(config, 'model_params.gpu_id', 1)
    U.set_nested(config, 'model_params.nthread', 12)
    U.set_nested(config, 'model_params.tree_method', 'hist')
    U.set_nested(config, 'fit_params.verbose', 100)
    U.set_nested(config, 'fit_params.early_stopping_rounds', 100)
    result = trainer.train(X_trn, features=features, fold=fold, config=config)
    ensemble.append(result)

rounders = [Rounder(method='dist'), Rounder(method='optimal')]
labels = []
for rounder in rounders:
    for i, result in enumerate(ensemble, 1):
        rounder.fit(X_trn['accuracy_group'], result.oof)
        for j, model in enumerate(result.models, 1):
            U.log(f'Rounder: {rounder.method}, ensemble #{i}, model #{j}')
            model_pred = model.predict(X_tst[features])
            model_labels = rounder.predict(model_pred)
            labels.append(model_labels)

bundle.save({
    'rounders': rounders, 
    'ensemble': [result.models for result in ensemble]
}, key='xgboost')

algo = 'lightgbm'
version = '024'

if U.on_kaggle():
    U.log('Inference on Kaggle.')
    features = bundle.features()
    bounds = bundle.bounds()
    predicted = inference(X_tst, features, bounds=bounds, model=algo, version=version)
    U.log('Saving predictions on disk.')
    filename = submit(predicted)
    submit_df = pd.read_csv(filename)
    U.log('First 20 submission rows:')
    display(submit_df.head(20))
    
else:
    U.log(f'Training model: {algo}')
    cappa = make_cappa_metric(X_trn['accuracy_group'])
    trainer = EnsembleTrainer(algo=algo, cv_metrics={'cappa': cappa})
    fold = GroupKFold(n_splits=5)
    config = get_default_config(algo)
    U.set_nested(config, 'model_params.feature_fraction', 0.4)
    U.set_nested(config, 'model_params.bagging_fraction', 0.5)
    U.set_nested(config, 'model_params.bagging_freq', 1)
    U.set_nested(config, 'model_params.learning_rate', 0.003)
    result = trainer.train(X_trn, features=features, fold=fold, config=config)
    U.log('Saving the trained models')
    bundle.save(result.models, f'models_{algo}_{version}')
    U.log('Saving the optimal rounding bounds')
    bounds = optimize_rounding_bounds(result.oof, X_trn['accuracy_group'].values)
    U.log(f'Optimal bounds: {bounds}')
    bundle.save(bounds, 'bounds')
    U.log(f'Final features number: {len(features)}')
    bundle.save(features, 'features')

if not U.on_kaggle():
    import os
    features = bundle.features()
    bounds = bundle.bounds()
    filename = submit(inference(X_tst, features, bounds, model=algo, version=version))
    assert os.path.exists(filename)
    assert pd.read_csv(filename).shape[0] == 1000
    bundle.package(folder='/home/ck/data/bowl2019/external/')

%matplotlib inline

pd.read_csv('submission.csv')['accuracy_group'].value_counts().reset_index().sort_values(by='index').set_index('index').plot.bar()

X_trn['accuracy_group'].value_counts().reset_index().sort_values(by='index').set_index('index').plot.bar()

## Stack Many Models

bundle.package(folder='/home/ck/data/bowl2019/external/')
